Split statement (;) support
This patch adds support for split statements by semicolons.  Unused
tokens are moved to the Lexer's cache, which are used in place of the
next line if not yet depleted.

The get_liminals() function was also modified to include semicolons.
This works because semicolons which appear at the beginning and end of
statements have no effect and act as liminals.

Note that we could use the cache in place of the lookahead iteration,
which could act as the lookahead and eliminate the need to tee().

The tester also removed the hard link to MOM.F90 and now expects a
command line input.