get_liminal now uses cache
We no longer need to use tee() to lookahead, and instead just move the
unparsed lexeme into the cache.  We then rely on the iterator to handle
the cache before reading the next line.

The biggest benefit here is that it eliminates the awkward empty-line
iterations caused by lines starting with semicolons.

This also simplifies a lot of the otherwise touchy tee-based methods.

We can probably start removing a lot of the prior_token conditional
checks, and possibly even remove the header variable, since we can
almost guarantee now that the leading lexeme is not liminal.

is_liminal was also re-defined to include the semicolon, and the one
remaining non-semicolon check just now explicitly looks for space and
comments.