prior_token replaced with prior_tail
The new cache-based approach now means that it's now always safe to
assume that the first lexeme is always semantic and never liminal, so we
no longer need to check if prior_tok exists.

Also it never needed to be a token!  So now it's prior_tail.

Some other useless stuff was removed (e.g. the is_lim lambda).