First attempt at line continuation
This is an initial implementation of line continuation handling.

We wrap the line iteration in an outer loop controlled by a line
continuation flag, which is immediately disabled upon entry.  When a
line continuation token is encountered, we immediately dump the
remaining tokens (assumed liminal) into the prior's tail, fetch the next
line, re-enable the flag, and then escape the loop.

This avoids a nested recursion, and instead just starts up the source
iteration from where it left off, but now the statement can be built
upon with the next line as if it is a new line.

This does not resolve the more complex cases where the subsequent line
has its own matching line continuation.  And of course no chance at all
of handling split tokens.  In fact, they are critically broken since
*all* leading & tokens will be skipped over.

But such usage is rare, and we know exactly how to resolve this case:
drop & from is_liminal and explictly look at the start of the loop when
the flag is true.

Overall this was much simpler than I expected, and this might be sorted
out much sooner than I expected.