Improved split line support
This fixes some of the issues with split lines, where any split was
moving the entire statement into the tail of the prior token.

We no longer attempt to update the statement in the current iteration,
and instead go through the generic lexeme update process, which has the
benefit of using cache in place of always pulling a new line.

This new approach also lets us lead with a get_liminal() call to pull
out liminals preceding any leading '&' tokens.

For now, we just dump the leading '&' and proceed as if there are no
other problems.  More work is needed to skillfully reconstruct any split
tokens.