Split token support
This is a first pass at split lexeme support.  Single splits are now
supported.  The original split token, with its whitespace, endlines,
splits, etc, is saved in the `split` field of the Token.

This appears to fail of a token is split more than once, but single
splits appear to work correctly.