Multiline support with & before/after breaks
Testing showed that our scanner was actually broken for string
continuation tests, and was hitting various problems when strings
containing & were split just before or after the &.

Somehow this got papered over as the old flint tokenizer would
reconstruct things, but the bottom line is that the tokens were flawed,
particularly for strings split just before &, e.g.

    s = "this &
        && is an ampersand"

The scanner now has a flag so that it has some partial awareness of
whether a token is a delimiter or something else (potentially part of a
string).  This change is not elegant, but the scanner needs a general
rewrite, and this patch is good enough to resolve the immediate problem.

The lexer was also modified to more safely reuse the self.scanner by
preserving self.scanner.prior_token when reusing it for small tasks
(currently just the split token test).

Finally, a testing function similar to lexer's was added to scanner.