Line continuation after & is liminal
We now handle the & token as if a liminal when following a semicolon
(statement separator) since it has no semantic meaning.  It was also
producing unexpected zero-length statements.

In following the new design, we immediately push all liminals into the
tail and make sure that every new iteration starts on a semantic token.