Stronger mulitline string splits
There were some odd issues with the string token splitting across
multiple lines.  In particular, we were not properly prepping the string
for the scanner.

We resolve these issues by conditionally appending & when working with a
split string.

(Probably a bit redundant, since we are saying that we already know if
this is a split token!  But works for now as a general test.)

Also:
Minor reformatting of the token test programs.