Partial preprocessing support
This patch adds some basic support for preprocessing:

* #define
* #undef
* #include

The #define/undef support is basically complete and works across
headers.  Substitutions are applied correctly.  At some point, we only
want to do this internally and track the original unsubstituted lexeme,
rather than explicitly altering it.

Header support via #include is working, but it only handles additional
preprocessing.  Actual Fortran statements are tokenized, but then
discarded once we leave the included header.

There is some handling of #ifdef/#else/#endif macro conditionals, but
the actual flag used to handle these is unused in the rest of the
Lexer.

Notably, the #if directive (including #if defined) is unhandled here,
since it would be difficult to parse the logical operations.

Finally, we only look in the current directory for headers, and make no
effort to traverse directories or even allow any way.

Despite those limitations, we are catching preprocessing, running it
through a dedicated preprocessor subroutine, then pushing the content
into the `tail` of the prior token.  Substitutions are applied.  More
work to be done, but this represents good progress.