Preprocessor macro support; token to ftoken
Much improved preprocessor macro support:
- We can now handle multiple lexemes inside of a macro's body, i.e. a
  macro can be substituted with multiple lexemes.

  Substitution was also reworked so that a PToken returns its redefined
  value when invoked (e.g. [lx for lx in lexemes]) but str() returns the
  original macro value.

Macros store themselves like this:
- The first lexeme stores the macro name in its preproc value (self.pp)
- The subsequent lexemes store empty strings in their self.pp
- The interior head/tail values are always empty strings.
- The edge ones just map to the adjacent values
  (e.g. ptok[0] = prior_tail)

I'll write up a small doc explaining this better.

Finally, the token.py module was renamed to ftoken.py since the original
name conflicted with token in the standard library.  The worst impact of
this was that help() was completely broken.  Now it works again.