Preprocessor macro conditional support
This patch adds support for macro conditional statements (#ifdef etc).
When macros disable a code block, all statements are treated as liminals
and stored in the "tails" of tokens.  When re-enabled, statements are
again scanned and lexed.

Some very skeletal code for parsing statements in include statements is
set up, but it's doubtful that it does anything close to what is
intended.